Add basic support of ps2dev/ps2sdk
URLs of the ps2dev/ps2sdk repos:
- https://github.com/ps2dev/ps2dev
- https://github.com/ps2dev/ps2sdk

+++ src/kyu/graphics/mesh.c
@@ -186,7 +186,7 @@ need_skip_line(int *index, const char *restrict line)
   *index = 0;
   skip_idx = 0;
       
-  while (line[*index] && isspace(line[*index]))
+  while (line[*index] && isspace((int)line[*index]))
     *index += 1;
 
   if (line[*index] == '\0')
